Fix regular org links
Add missing snippet of code to enable dynamic links to work. However, it
looks like all the other type of links use the `htmlize-link' property.
And so when you click on a file link, browse-url tries to open it and
fail. So now for dynamic links, we use a different property:
`activity-log-link'.

After this change, both links work fine.